Deleted wrong casting

+++ src/Native/Runtime/unix/UnixContext.cpp
@@ -208,11 +208,11 @@ static void RegDisplayToUnwindContext(REGDISPLAY* regDisplay, unw_context_t *unw
     // unw_set_reg().
 
 #define ASSIGN_REG(regIndex, regName)                           \
-    unwContext->data[regIndex] = (uint64_t)(regDisplay->regName);
+    unwContext->data[regIndex] = (regDisplay->regName);
 
 #define ASSIGN_REG_PTR(regIndex, regName) \
     if (regDisplay->p##regName != NULL) \
-        unwContext->data[regIndex] = (uint64_t)*(regDisplay->p##regName);
+        unwContext->data[regIndex] = *(regDisplay->p##regName);
 
     ASSIGN_REG_PTR(4, R4);
     ASSIGN_REG_PTR(5, R5);
